scale singular values for tolerance
The chemical rank tolerance was meaningless unless the singular values
were normalized to the maximum singular value to give a fractional
significance. Therefore the renamed chemical rank tolerance should
be understood as a fraction of the most significant component in
terms of influence, and one may impose something like 5% or 1%
significance (since some intermediate species may be very low in
concentration). Alternatives are to scale by the sum of all singular
values.

Also did not reassign the data matrix variable name A but gave distinct
variable names A, B, C, D to transforms. Since the transforms are not
in-place this is ideal for debugging since it allows the data which is still in
memory to be accessed.